further divide EntityFetcher
